# contracts/openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: Knowledge Platform API
  version: 0.1.0
  description: |
    Canonical HTTP contract for the independently implemented knowledge platform.
    All identity fields named "email" accept an email address only; usernames do not exist.
servers:
  - url: /api/v1
...
paths:
...
  /auth/login/email-code/request:
    post:
      tags: [Authentication]
      operationId: requestEmailLoginCode
      description: Always returns 202 to prevent account enumeration.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailCodeRequest"
      responses:
        "202":
          description: Accepted whether or not the account exists.
  /auth/login/email-code/verify:
    post:
      tags: [Authentication]
      operationId: loginWithEmailCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailCodeLoginRequest"
      responses:
        "204":
          description: Authenticated; the response sets a secure session cookie.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KP_SESSION
  schemas:
...
    EmailCodeRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
    EmailCodeLoginRequest:
      type: object
      additionalProperties: false
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: "^[0-9]{6}$"
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
  responses:
...
    Unauthorized:
      description: Credentials are invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
security:
  - sessionCookie: []